requirements.txt
gymnasium>=0.29.0
stable-baselines3>=2.3.0
pyyaml>=6.0
diskcache>=5.6.0
matplotlib>=3.8.0
seaborn>=0.13.0
numpy>=1.26.0
pytest>=8.0.0
